tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.initial.reset.clip.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.initial.reset.clip</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.initial.reset.clip</h1>
<p class="desc">Resetting the canvas state resets the current clip region</p>


<script>
var t = async_test("Resetting the canvas state resets the current clip region");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.width = 100;
  ctx.rect(0, 0, 1, 1);
  ctx.clip();
  canvas.width = 100;
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 20,20, 0,255,0,255);
  t.done();

});
</script>
